requirements.txt
Django==1.6
django-leaflet==0.15.2
jsonfield==1.0.1
mongoengine==0.8.7
pymongo==2.8
six==1.9.0
